# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRefinementRequest": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetConstraint-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "constraints": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Constraints",
            "type": "array"
          },
          "instruction": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          },
          "preferences": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Preferences",
            "type": "array"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionPreference-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationMode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "instruction"
        ],
        "title": "CreateRefinementRequest",
        "type": "object"
      },
...
      "CreateSessionRequest-Output": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetConstraint-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Constraints",
            "type": "array"
          },
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Preferences",
            "type": "array"
          },
          "query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionPreference-Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query"
        ],
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "CreateUserAddedProductRequest": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "fallback_candidate_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Candidate Id"
          },
          "input_text": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Text"
          },
          "manual_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualProductDetails-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "manual_fallback_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualFallbackReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "CreateUserAddedProductRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualFallbackReason": {
        "enum": [
          "retrieval_unavailable",
          "retrieval_insufficient",
          "user_correction"
        ],
        "title": "ManualFallbackReason",
        "type": "string"
      },
      "ManualProductDetails-Input": {
        "additionalProperties": false,
        "description": "Shopper reports, never independently verified product or offer evidence.",
        "properties": {
          "availability": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability"
          },
          "price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "review": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review"
          },
          "seller": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller"
          },
          "specifications": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specifications"
          },
          "warranty": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warranty"
          }
        },
        "title": "ManualProductDetails",
        "type": "object"
      },
...
      "RecomputePlan": {
        "additionalProperties": false,
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactDecision"
            },
            "minItems": 1,
            "title": "Artifacts",
            "type": "array"
          },
          "base_brief": {
            "$ref": "#/components/schemas/ShoppingBrief"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "prior_result_version_id": {
            "format": "uuid",
            "title": "Prior Result Version Id",
            "type": "string"
          },
          "prior_run_id": {
            "format": "uuid",
            "title": "Prior Run Id",
            "type": "string"
          },
          "refinement_id": {
            "format": "uuid",
            "title": "Refinement Id",
            "type": "string"
          },
          "requested_result_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/RecomputeStage"
            },
            "minItems": 1,
            "title": "Stages",
            "type": "array"
          },
          "target_brief": {
            "$ref": "#/components/schemas/ShoppingBrief"
          }
        },
        "required": [
          "refinement_id",
          "session_id",
          "run_id",
          "prior_run_id",
          "prior_result_version_id",
          "base_brief",
          "target_brief",
          "stages",
          "artifacts"
        ],
        "title": "RecomputePlan",
        "type": "object"
      },
...
      "RefinementRunResponse": {
        "additionalProperties": false,
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/RecomputePlan"
          },
          "refinement": {
            "$ref": "#/components/schemas/RefinementRequest"
          },
          "run": {
            "$ref": "#/components/schemas/ShoppingRunRecord"
          }
        },
        "required": [
          "refinement",
          "run",
          "plan"
        ],
        "title": "RefinementRunResponse",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "SessionStateResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_brief": {
            "$ref": "#/components/schemas/ShoppingBrief"
          },
          "original_input": {
            "$ref": "#/components/schemas/CreateSessionRequest-Output"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_added_products": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UserAddedProduct"
            },
            "title": "User Added Products",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "session_id",
          "original_input",
          "current_brief",
          "created_at",
          "updated_at"
        ],
        "title": "SessionStateResponse",
        "type": "object"
      },
      "ShoppingBrief": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetConstraint-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "category_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FieldSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Constraints",
            "type": "array"
          },
          "original_query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Original Query",
            "type": "string"
          },
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/PreferenceConstraint"
            },
            "title": "Preferences",
            "type": "array"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionPreference-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "original_query"
        ],
        "title": "ShoppingBrief",
        "type": "object"
      },
...
      "ShoppingRunRecord": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "pending"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ShoppingRunRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CartCart Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions/{session_id}/products": {
      "post": {
        "operationId": "add_user_added_product_api_sessions__session_id__products_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAddedProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add User Added Product",
        "tags": [
          "products"
        ]
      }
    },
    "/api/sessions/{session_id}/refinements": {
      "post": {
        "operationId": "create_refinement_api_sessions__session_id__refinements_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefinementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefinementRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Refinement",
        "tags": [
          "refinements"
        ]
      }
    },
    "/api/sessions/{session_id}/refinements/{refinement_id}/execute": {
      "post": {
        "operationId": "execute_refinement_api_sessions__session_id__refinements__refinement_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "refinement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Refinement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingRunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Refinement",
        "tags": [
          "refinements"
        ]
      }
    },
    "/api/sessions/{session_id}/refinements/{refinement_id}/plan": {
      "get": {
        "operationId": "get_refinement_plan_api_sessions__session_id__refinements__refinement_id__plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "refinement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Refinement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecomputePlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Refinement Plan",
        "tags": [
          "refinements"
        ]
      }
    },
...
  }
}